+++ lib/src/messages/partial_custom.dart
@@ -1,5 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
+
 import 'custom_message.dart';
 
 part 'partial_custom.g.dart';