+++ lib/src/messages/partial_file.dart
@@ -1,5 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
+
 import 'file_message.dart';
 
 part 'partial_file.g.dart';