+++ lib/src/messages/partial_image.dart
@@ -1,5 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
+
 import 'image_message.dart';
 
 part 'partial_image.g.dart';