+++ lib/src/messages/partial_text.dart
@@ -1,5 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
+
 import '../preview_data.dart' show PreviewData;
 import 'text_message.dart';